Create Alternate Additions
Problem: Chef has two numbers, A and B, where A is less than B. Chef can perform operations on A by adding either 1 or 2 to it. The goal is for Chef to make A equal to B.

Solution: Chef can add 1 to A if A is odd, and add 2 to A if A is even. He can repeat these operations until A is equal to B.