Create Chef and digits of a number.cpp
The problem involves Chef trying to modify a binary number by flipping exactly one digit to make all the digits equal (either all 0's or all 1's). The input consists of the number of test cases (T) followed by T lines, each containing a binary number. The task is to determine if it's possible to achieve the desired outcome and output "Yes" if possible, and "No" otherwise.